Don't group the two classes in the same header file
