+++ src/revlanguage/workspace/RbRegister_Func.cpp
@@ -265,11 +265,13 @@
 #include "Func_logistic.h"
 #include "Func_matrix.h"
 #include "Func_max.h"
+#include "Func_maxMatrix.h"
 #include "Func_mean.h"
 #include "Func_meanPositive.h"
 #include "Func_meanSimplex.h"
 #include "Func_median.h"
 #include "Func_min.h"
+#include "Func_minMatrix.h"
 #include "Func_normalize.h"
 #include "Func_posteriorPredictiveProbability.h"
 //#include "Func_power.h"
